refactor(decks): reorganize view components into logical subfolders
