+++ Features/Decks/Views/Components/ActionLibrary/ActionLibraryView.xaml
@@ -1,15 +1,12 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.ActionLibraryView"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.ActionLibrary.ActionLibraryView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
              xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
              xmlns:dd="clr-namespace:CustomDragDropLib;assembly=CustomDragDropLib"
-             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components"
+             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components.ActionLibrary"
+             xmlns:layout="clr-namespace:StreamBoard.Features.Decks.Views.Components.Layout"
              xmlns:controls="clr-namespace:StreamBoard.Components.Controls"
-             xmlns:vm="clr-namespace:StreamBoard.Features.Decks.ViewModels"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
+             xmlns:vm="clr-namespace:StreamBoard.Features.Decks.ViewModels">
 
     <UserControl.Resources>
         <DataTemplate x:Key="ActionGhostTemplate">
@@ -36,8 +33,8 @@
             <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
 
-        <local:DeckSidebarHeader Title="Action Library">
-            <local:DeckSidebarHeader.BottomContent>
+        <layout:DeckSidebarHeader Title="Action Library">
+            <layout:DeckSidebarHeader.BottomContent>
                 <ui:TextBox PlaceholderText="Search actions..." 
                             Text="{Binding SearchText, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
                             Margin="8,0,8,8"
@@ -46,8 +43,8 @@
                         <controls:FluentIcon IconType="Search" />
                     </ui:TextBox.Icon>
                 </ui:TextBox>
-            </local:DeckSidebarHeader.BottomContent>
-        </local:DeckSidebarHeader>
+            </layout:DeckSidebarHeader.BottomContent>
+        </layout:DeckSidebarHeader>
 
         <ScrollViewer Grid.Row="1"
                       VerticalScrollBarVisibility="Auto"
