+++ Features/Decks/Views/Components/ActionLibrary/ActionLibraryView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Controls;
 
-namespace StreamBoard.Features.Decks.Views.Components
+namespace StreamBoard.Features.Decks.Views.Components.ActionLibrary
 {
     public partial class ActionLibraryView : UserControl
     {
