+++ Features/Decks/Views/Components/Canvas/DeckButtonSlotView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.DeckButtonSlotView"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.Canvas.DeckButtonSlotView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
