+++ Features/Decks/Views/Components/Canvas/GridCanvasView.xaml
@@ -1,7 +1,7 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.GridCanvasView"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.Canvas.GridCanvasView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components"
+             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components.Canvas"
              xmlns:models="clr-namespace:StreamBoard.Features.Decks.Models"
              xmlns:dd="clr-namespace:CustomDragDropLib;assembly=CustomDragDropLib"
              x:Name="Root">
