+++ Features/Decks/Views/Components/Canvas/GridCanvasView.xaml.cs
@@ -2,7 +2,7 @@
 using System.Windows.Controls;
 using System.Windows.Input;
 
-namespace StreamBoard.Features.Decks.Views.Components
+namespace StreamBoard.Features.Decks.Views.Components.Canvas
 {
     public partial class GridCanvasView : UserControl
     {
