+++ Features/Decks/Views/Components/DeckProfile/DeckPageItem.xaml.cs
@@ -1,11 +1,10 @@
 ﻿using StreamBoard.Features.Decks.ViewModels;
-using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Media;
 
-namespace StreamBoard.Features.Decks.Views.Components
+namespace StreamBoard.Features.Decks.Views.Components.DeckProfile
 {
     public partial class DeckPageItem : UserControl
     {
