+++ Features/Decks/Views/Components/DeckProfile/DeckProfilyView.xaml
@@ -1,10 +1,11 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.DeckProfilyView"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.DeckProfile.DeckProfilyView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
              xmlns:dd="clr-namespace:GongSolutions.Wpf.DragDrop;assembly=GongSolutions.Wpf.DragDrop"
-             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components"
+             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components.DeckProfile"
+             xmlns:layout="clr-namespace:StreamBoard.Features.Decks.Views.Components.Layout"
              xmlns:controls="clr-namespace:StreamBoard.Components.Controls">
 
     <UserControl.Resources>
@@ -27,15 +28,15 @@
             <RowDefinition Height="*" />
         </Grid.RowDefinitions>
 
-        <local:DeckSidebarHeader Grid.Row="0" Title="Pages">
-            <local:DeckSidebarHeader.RightContent>
+        <layout:DeckSidebarHeader Grid.Row="0" Title="Pages">
+            <layout:DeckSidebarHeader.RightContent>
                 <StackPanel Orientation="Horizontal">
                     <controls:IconButton TooltipText="Add New Page" Command="{Binding AddPageCommand}" IconType="Add" />                    
                     <controls:IconButton TooltipText="Rename Page" Command="{Binding RenamePageCommand}" IconType="Rename" Margin="8,0,0,0"/>
                     <controls:IconButton TooltipText="Delete Page" Command="{Binding DeletePageCommand}" IconType="Delete" Margin="8,0,0,0"/>
                 </StackPanel>
-            </local:DeckSidebarHeader.RightContent>
-        </local:DeckSidebarHeader>
+            </layout:DeckSidebarHeader.RightContent>
+        </layout:DeckSidebarHeader>
 
         <ScrollViewer Grid.Row="1" 
                   VerticalScrollBarVisibility="Auto"
