+++ Features/Decks/Views/Components/DeckProfile/DeckProfilyView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Controls;
 
-namespace StreamBoard.Features.Decks.Views.Components
+namespace StreamBoard.Features.Decks.Views.Components.DeckProfile
 {
     public partial class DeckProfilyView : UserControl
     {
