+++ Features/Decks/Views/Components/Layout/DeckEditorLayout.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.DeckEditorLayout"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.Layout.DeckEditorLayout"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
