+++ Features/Decks/Views/Components/Layout/DeckEditorLayout.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows;
 using System.Windows.Controls;
 
-namespace StreamBoard.Features.Decks.Views.Components
+namespace StreamBoard.Features.Decks.Views.Components.Layout
 {
     public partial class DeckEditorLayout : UserControl
     {
