+++ Features/Decks/Views/Components/PropertyEditor/ButtonActionItem.xaml
@@ -1,7 +1,6 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.ButtonActionItem"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.PropertyEditor.ButtonActionItem"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components"
              xmlns:controls="clr-namespace:StreamBoard.Components.Controls"
              x:Name="Root">
