+++ Features/Decks/Views/Components/PropertyEditor/ImagePicker.xaml
@@ -1,7 +1,6 @@
-﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.ImagePicker"
+﻿<UserControl x:Class="StreamBoard.Features.Decks.Views.Components.PropertyEditor.ImagePicker"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Components"
              xmlns:controls="clr-namespace:StreamBoard.Components.Controls">
     
     <Grid Background="Transparent"
