+++ Features/Decks/Views/Pages/GridDeckPage.xaml
@@ -4,29 +4,33 @@
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
       xmlns:local="clr-namespace:StreamBoard.Features.Decks.Views.Pages"
-      xmlns:components="clr-namespace:StreamBoard.Features.Decks.Views.Components"
+      xmlns:layout="clr-namespace:StreamBoard.Features.Decks.Views.Components.Layout"
+      xmlns:deck_profile="clr-namespace:StreamBoard.Features.Decks.Views.Components.DeckProfile"
+      xmlns:canvas="clr-namespace:StreamBoard.Features.Decks.Views.Components.Canvas"
+      xmlns:action_library="clr-namespace:StreamBoard.Features.Decks.Views.Components.ActionLibrary"
+      xmlns:property_editor="clr-namespace:StreamBoard.Features.Decks.Views.Components.PropertyEditor"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
       Title="GridDeckPage"
       ScrollViewer.CanContentScroll="False">
 
-    <components:DeckEditorLayout VerticalAlignment="Stretch" HorizontalAlignment="Stretch">
+    <layout:DeckEditorLayout VerticalAlignment="Stretch" HorizontalAlignment="Stretch">
 
-        <components:DeckEditorLayout.LeftBarContent>
-            <components:DeckProfilyView DataContext="{Binding Pages}"/>
-        </components:DeckEditorLayout.LeftBarContent>
+        <layout:DeckEditorLayout.LeftBarContent>
+            <deck_profile:DeckProfilyView DataContext="{Binding Pages}"/>
+        </layout:DeckEditorLayout.LeftBarContent>
 
-        <components:DeckEditorLayout.CanvasContent>
-            <components:GridCanvasView DataContext="{Binding}"/>
-        </components:DeckEditorLayout.CanvasContent>
+        <layout:DeckEditorLayout.CanvasContent>
+            <canvas:GridCanvasView DataContext="{Binding}"/>
+        </layout:DeckEditorLayout.CanvasContent>
 
-        <components:DeckEditorLayout.PropertiesContent>
-            <components:PropertyEditorView DataContext="{Binding}"/>
-        </components:DeckEditorLayout.PropertiesContent>
+        <layout:DeckEditorLayout.PropertiesContent>
+            <property_editor:PropertyEditorView DataContext="{Binding}"/>
+        </layout:DeckEditorLayout.PropertiesContent>
 
-        <components:DeckEditorLayout.LibraryContent>
-            <components:ActionLibraryView DataContext="{Binding Library}" />
-        </components:DeckEditorLayout.LibraryContent>
+        <layout:DeckEditorLayout.LibraryContent>
+            <action_library:ActionLibraryView DataContext="{Binding Library}" />
+        </layout:DeckEditorLayout.LibraryContent>
 
-    </components:DeckEditorLayout>
+    </layout:DeckEditorLayout>
 </Page>
